Added a dry run feature where you can check a box or supply a --dryRun parameter and it will show what it would do if the flag wasn't set, but doesn't actually do anything. It adds "[Dry Run]" in front the the log messages that are not actually done.

+++ JamfSync/Model/JamfProInstance.swift
@@ -151,14 +151,16 @@ class JamfProInstance: SavableItem {
     /// - Parameters:
     ///     - srcDp: The destination distribution point to search and delete packages that are missing.
     ///     - progress: The progress object that should be updated as the deletion progresses.
-    func deletePackagesNotOnSource(srcDp: DistributionPoint, progress: SynchronizationProgress) async throws {
+    func deletePackagesNotOnSource(srcDp: DistributionPoint, progress: SynchronizationProgress, dryRun: Bool) async throws {
         guard let packageApi else { throw DistributionPointError.programError }
         let packagesToRemove = packagesToRemove(srcDp: srcDp)
         for package in packagesToRemove {
             if let jamfProId = package.jamfProId {
-                LogManager.shared.logMessage(message: "Deleting package \(package.fileName) from \(displayName())", level: .verbose)
-                try await packageApi.deletePackage(packageId: jamfProId, jamfProInstance: self)
-                packages.removeAll(where: { $0.fileName == package.fileName })
+                LogManager.shared.logMessage(message: "Deleting package \(package.fileName) from \(displayName())", level: .verbose, dryRun: dryRun)
+                if !dryRun {
+                    try await packageApi.deletePackage(packageId: jamfProId, jamfProInstance: self)
+                    packages.removeAll(where: { $0.fileName == package.fileName })
+                }
             }
         }
     }
